TR004_input implementation code
